+++ Mikibot.Crawler/WebsocketCrawler/Data/Commands/KnownCommand/DanmuMsg.cs
@@ -11,11 +11,11 @@ namespace Mikibot.Crawler.WebsocketCrawler.Data.Commands.KnownCommand
     public struct DanmuMsg
     {
         public string Msg { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public string UserName { get; set; }
         public string FansTag { get; set; }
         public int FansLevel { get; set; }
-        public int FansTagUserId { get; set; }
+        public long FansTagUserId { get; set; }
         public string FansTagUserName { get; set; }
         public DateTimeOffset SentAt { get; set; }
         public string MemeUrl { get; set; }
