+++ Mikibot.Crawler/WebsocketCrawler/Data/Commands/KnownCommand/InteractFansMedal.cs
@@ -16,8 +16,8 @@ namespace Mikibot.Crawler.WebsocketCrawler.Data.Commands.KnownCommand
         [JsonPropertyName("medal_name")]
         public string MedalName { get; set; }
         [JsonPropertyName("target_id")]
-        public int FansTagUserId { get; set; }
+        public long FansTagUserId { get; set; }
         [JsonPropertyName("anchor_roomid")]
-        public int FansTagRoomId { get; set; }
+        public long FansTagRoomId { get; set; }
     }
 }
